Fix: Icon New , Unchanged & remove case handled

+++ app/Services/GenericSliderService.php
@@ -147,7 +147,7 @@ class GenericSliderService
                 Redis::del('non_bl_offer');
             }
 
-            if (isset($data['icon']) && $data['icon'] != 'not-updated') {
+            if (isset($data['icon']) && $data['icon'] != 'not-updated' && $data['icon'] != 'removed') {
                 $data['icon'] = 'storage/' . $data['icon']->store('generic_sliders_icons');
                 if (isset($slider) && file_exists($slider->icon)) {
                     unlink($slider->icon);
